scaffolder-backend: move more types to -node
