+++ .changeset/honest-fireants-applaud.md
@@ -1,0 +1,17 @@
+---
+'@backstage/plugin-scaffolder-backend': patch
+---
+
+Deprecated the following type exports, which have been moved to `@backstage/plugin-scaffolder-node` instead:
+
+- `TemplateFilter`
+- `TemplateGlobal`
+- `TaskStatus`
+- `TaskCompletionState`
+- `SerializedTask`
+- `TaskEventType`
+- `SerializedTaskEvent`
+- `TaskBrokerDispatchResult`
+- `TaskBrokerDispatchOptions`
+- `TaskContext`
+- `TaskBroker`
