+++ .changeset/little-peas-decide.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-scaffolder-node': patch
+---
+
+Added several new types that were moved from `@backstage/plugin-scaffolder-backend`.
