+++ plugins/scaffolder-backend/src/scaffolder/tasks/types.ts
@@ -14,131 +14,94 @@
  * limitations under the License.
  */
 
-import { JsonValue, JsonObject, Observable } from '@backstage/types';
+import { JsonValue, JsonObject } from '@backstage/types';
 import { TaskSpec, TaskStep } from '@backstage/plugin-scaffolder-common';
 import { TaskSecrets } from '@backstage/plugin-scaffolder-node';
-import { TemplateAction } from '@backstage/plugin-scaffolder-node';
+import {
+  TemplateAction,
+  TaskStatus as _TaskStatus,
+  TaskCompletionState as _TaskCompletionState,
+  SerializedTask as _SerializedTask,
+  TaskEventType as _TaskEventType,
+  SerializedTaskEvent as _SerializedTaskEvent,
+  TaskBrokerDispatchResult as _TaskBrokerDispatchResult,
+  TaskBrokerDispatchOptions as _TaskBrokerDispatchOptions,
+  TaskContext as _TaskContext,
+  TaskBroker as _TaskBroker,
+} from '@backstage/plugin-scaffolder-node';
 
 /**
  * The status of each step of the Task
  *
  * @public
+ * @deprecated Import from `@backstage/plugin-scaffolder-node` instead.
  */
-export type TaskStatus =
-  | 'cancelled'
-  | 'completed'
-  | 'failed'
-  | 'open'
-  | 'processing';
+export type TaskStatus = _TaskStatus;
 
 /**
  * The state of a completed task.
  *
  * @public
+ * @deprecated Import from `@backstage/plugin-scaffolder-node` instead.
  */
-export type TaskCompletionState = 'failed' | 'completed';
+export type TaskCompletionState = _TaskCompletionState;
 
 /**
  * SerializedTask
  *
  * @public
+ * @deprecated Import from `@backstage/plugin-scaffolder-node` instead.
  */
-export type SerializedTask = {
-  id: string;
-  spec: TaskSpec;
-  status: TaskStatus;
-  createdAt: string;
-  lastHeartbeatAt?: string;
-  createdBy?: string;
-  secrets?: TaskSecrets;
-};
+export type SerializedTask = _SerializedTask;
 
 /**
  * TaskEventType
  *
  * @public
+ * @deprecated Import from `@backstage/plugin-scaffolder-node` instead.
  */
-export type TaskEventType = 'completion' | 'log' | 'cancelled';
+export type TaskEventType = _TaskEventType;
 
 /**
  * SerializedTaskEvent
  *
  * @public
+ * @deprecated Import from `@backstage/plugin-scaffolder-node` instead.
  */
-export type SerializedTaskEvent = {
-  id: number;
-  taskId: string;
-  body: JsonObject;
-  type: TaskEventType;
-  createdAt: string;
-};
+export type SerializedTaskEvent = _SerializedTaskEvent;
 
 /**
  * The result of {@link TaskBroker.dispatch}
  *
  * @public
+ * @deprecated Import from `@backstage/plugin-scaffolder-node` instead.
  */
-export type TaskBrokerDispatchResult = {
-  taskId: string;
-};
+export type TaskBrokerDispatchResult = _TaskBrokerDispatchResult;
 
 /**
  * The options passed to {@link TaskBroker.dispatch}
  * Currently a spec and optional secrets
  *
  * @public
+ * @deprecated Import from `@backstage/plugin-scaffolder-node` instead.
  */
-export type TaskBrokerDispatchOptions = {
-  spec: TaskSpec;
-  secrets?: TaskSecrets;
-  createdBy?: string;
-};
+export type TaskBrokerDispatchOptions = _TaskBrokerDispatchOptions;
 
 /**
  * Task
  *
  * @public
+ * @deprecated Import from `@backstage/plugin-scaffolder-node` instead.
  */
-export interface TaskContext {
-  cancelSignal: AbortSignal;
-  spec: TaskSpec;
-  secrets?: TaskSecrets;
-  createdBy?: string;
-  done: boolean;
-  isDryRun?: boolean;
-
-  complete(result: TaskCompletionState, metadata?: JsonObject): Promise<void>;
-
-  emitLog(message: string, logMetadata?: JsonObject): Promise<void>;
-
-  getWorkspaceName(): Promise<string>;
-}
+export type TaskContext = _TaskContext;
 
 /**
  * TaskBroker
  *
  * @public
+ * @deprecated Import from `@backstage/plugin-scaffolder-node` instead.
  */
-export interface TaskBroker {
-  cancel?(taskId: string): Promise<void>;
-
-  claim(): Promise<TaskContext>;
-
-  dispatch(
-    options: TaskBrokerDispatchOptions,
-  ): Promise<TaskBrokerDispatchResult>;
-
-  vacuumTasks(options: { timeoutS: number }): Promise<void>;
-
-  event$(options: {
-    taskId: string;
-    after: number | undefined;
-  }): Observable<{ events: SerializedTaskEvent[] }>;
-
-  get(taskId: string): Promise<SerializedTask>;
-
-  list?(options?: { createdBy?: string }): Promise<{ tasks: SerializedTask[] }>;
-}
+export type TaskBroker = _TaskBroker;
 
 /**
  * TaskStoreEmitOptions
