+++ plugins/scaffolder-node/src/types.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Backstage Authors
+ * Copyright 2023 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,16 +14,12 @@
  * limitations under the License.
  */
 
-/**
- * The scaffolder-node module for `@backstage/plugin-scaffolder-backend`.
- *
- * @packageDocumentation
- */
+import { JsonValue } from '@backstage/types';
 
-export * from './actions';
-export * from './tasks';
-export type { TemplateFilter, TemplateGlobal } from './types';
-export {
-  scaffolderActionsExtensionPoint,
-  type ScaffolderActionsExtensionPoint,
-} from './extensions';
+/** @public */
+export type TemplateFilter = (...args: JsonValue[]) => JsonValue | undefined;
+
+/** @public */
+export type TemplateGlobal =
+  | ((...args: JsonValue[]) => JsonValue | undefined)
+  | JsonValue;
